requirements.txt
Flask==1.0.2
google-cloud-datastore==1.7.1
google-auth==1.6.1
anom==0.9.1
google-cloud-storage==1.13.0
google-cloud-logging==1.8.0
gunicorn==19.9.0
oauth2client==4.1.3
six==1.11.0
requests[security]==2.20.1
mock==2.0.0
google-cloud-firestore==0.31.0
